+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `tt cluster replicaset roles add`: command to add roles in config scope provided by flags.
+- `tt replicaset roles remove`: command to remove roles in the tarantool replicaset with cluster
+  config (3.0) or cartridge orchestrator.
 
 ### Fixed